@@ -7,10 +7,12 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
-    "react": "^18.0.0",
+    "ajv": "^8.11.0",
+    "axios": "^0.27.2",
+    "react": "^18.2.0",
     "react-autocomplete": "^1.8.1",
     "react-bootstrap": "^2.2.3",
-    "react-dom": "^18.0.0",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.29.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.0",
