@@ -1,23 +1,31 @@
 import PropTypes from "prop-types";
 import Col from "react-bootstrap/Col";
 import Card from "react-bootstrap/Card";
+import Button from "react-bootstrap/esm/Button";
+
 
 function MonsterCard({ id, name, imageUrl, types, hp, artist, rarity }) {
 	const { content } = ({ id, name, imageUrl, types, hp, artist, rarity })
-    return (
+    
+	return (
+		
 		<Col>
 			
+			
 				<Card style={{ width: "18rem" }}>
+					<Card.Title>{name}</Card.Title>
 					<Card.Img variant="top" src={imageUrl} />
-					<Card.Body>
-						<Card.Title>{name}</Card.Title>
+					<Card.Body>						
 						<div>
-							{content}
-						</div>
+							{content}							
+						</div>	
+						<Button>Details Page</Button>
+																	
 					</Card.Body>
 				</Card>
 			
 		</Col>
+		
 	);
 }
 
